+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 - Move to Java 17.
 - Add base functionality to apply fixes from magik-lint.
+- Add UnsafeEvaluateInvocationCheck to test for unsafe `unsafe_evaluate()` method calls.
+- Add SwChar16VectorEvaluateInvocationCheck to test for `sw:char16_vector.evaluate()` method calls.
 
 0.9.0 (2024-02-26)